pyproject.toml
[tool.poetry]
name = "ml-mesh"
version = "0.1.0"
description = "Development dependencies"
authors = ["Rene Jean Corneille <rene-jean.corneille@onclusive.com>"]

[tool.poetry.dependencies]
python = "3.8.16"
click = "8.1.0"
boto3 = "1.26.58"
core = {path = "libs/core"}
pytest = "7.2.0"
python-dotenv = "1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
